[not verified] refactoring -> add a new use-share-post hook
