+++ projects/plugins/jetpack/extensions/plugins/publicize/hooks/use-share-post/Readme.md
@@ -1,0 +1,3 @@
+# useSharePost() hook
+
+React hook to share a post to social media connections
